Extract the cranelift compile-time evaluator into a submodule
Second cranelift_backend.rs split under #1254: the compile-time program
evaluator -- the SpikeValue interpreter and its host-capability call
dispatchers (fs, net, http, async, process, clock, crypto, json/serdes,
regex, encoding), 148 functions -- moves to
stage1/crates/axiomc/src/cranelift_backend/evaluator.rs, a sibling of
the intrinsics module. cranelift_backend.rs drops 27,815 -> 23,700
lines and the top-file share falls below 0.78.

Pure code motion. Shared value types (SpikeValue and companions) and
the helpers reused by the i64 lowering path stay in the parent module;
the submodule sees them through `use super::*`, and the types that
appear in moved pub(crate) signatures are raised to pub(crate) to keep
the private-interfaces lint clean. Call sites are unchanged via
`pub(crate) use evaluator::*`.

The plan records the lowered cranelift ceiling and the new evaluator
ceiling.

Refs #1254.
